+++ biopipen/scripts/scrna/TopExpressingGenes.R
@@ -12,18 +12,18 @@ srtfile <- {{in.srtobj | r}}
 outdir <- {{out.outdir | r}}
 mutaters <- {{ envs.mutaters | r }}
 ident <- {{ envs.ident | r }}
-group.by <- {{ envs["group-by"] | r }}
+group.by <- {{ envs["group-by"] | r }}  # nolint
 each <- {{ envs.each | r }}
 prefix_each <- {{ envs.prefix_each | r }}
 section <- {{ envs.section | r }}
 dbs <- {{ envs.dbs | r }}
 n <- {{ envs.n | r }}
-cases <- {{ envs.cases | r: todot="-" }}
+cases <- {{ envs.cases | r: todot = "-" }}  # nolint
 
 set.seed(8525)
 
 print("- Loading Seurat object ...")
-srtobj = readRDS(srtfile)
+srtobj <- readRDS(srtfile)
 
 print("- Mutate meta data if needed ...")
 if (!is.null(mutaters) && length(mutaters)) {
@@ -33,8 +33,8 @@ if (!is.null(mutaters) && length(mutaters)) {
 
 print("- Expanding cases ...")
 if (is.null(cases) || length(cases) == 0) {
-    cases = list(
-        Cluster = list(
+    cases <- list(
+        DEFAULT = list(
             ident = ident,
             group.by = group.by,
             each = each,
@@ -45,7 +45,7 @@ if (is.null(cases) || length(cases) == 0) {
         )
     )
 } else {
-    cases = lapply(cases, function(cs) {
+    cases <- lapply(cases, function(cs) {
         list_setdefault(
             cs,
             ident = ident,
@@ -61,14 +61,17 @@ if (is.null(cases) || length(cases) == 0) {
 
 # Expand each and ident
 newcases <- list()
-for (name in names(cases)) {
+for (name in names(cases)) {  # nolint
     case <- cases[[name]]
     if (is.null(case$each) && !is.null(case$ident)) {
         newcases[[paste0(case$section, ":", name)]] <- case
     } else if (is.null(case$each)) {
-        idents <- srtobj@meta.data %>% pull(case$group.by) %>% unique() %>% na.omit()
+        idents <- srtobj@meta.data %>%
+            pull(case$group.by) %>%
+            unique() %>%
+            na.omit()
         for (ident in idents) {
-            key = paste0(name, ":", ident)
+            key <- paste0(name, ":", ident)
             newcases[[key]] <- case
             newcases[[key]]$ident <- ident
         }
@@ -76,27 +79,33 @@ for (name in names(cases)) {
         eachs <- srtobj@meta.data %>% pull(case$each) %>% unique() %>% na.omit()
         for (each in eachs) {
             by <- make.names(paste0(".", name, "_", each))
-            srtobj@meta.data = srtobj@meta.data %>% mutate(
+            srtobj@meta.data <- srtobj@meta.data %>% mutate(
                 !!sym(by) := if_else(
                     !!sym(case$each) == each,
                     !!sym(case$group.by),
                     NA
                 )
             )
-            case$group.by <- by
             if (is.null(case$ident)) {
-                idents <- srtobj@meta.data %>% pull(case$group.by) %>% unique() %>% na.omit()
+                idents <- srtobj@meta.data %>%
+                    pull(case$group.by) %>%
+                    unique() %>%
+                    na.omit()
                 for (ident in idents) {
-                    key = if (case$prefix_each) {
-                        paste0(case$each, "-", each, "-", name, ":", ident)
-                    } else {
-                        paste0(each, "-", name, ":", ident)
+                    kname <- if (name == "DEFAULT") "" else paste0("-", name)
+                    key <- paste0(each, kname, ":", ident)
+                    if (case$prefix_each) {
+                        key <- paste0(case$each, "-", key)
                     }
                     newcases[[key]] <- case
                     newcases[[key]]$ident <- ident
+                    newcases[[key]]$group.by <- by  # nolint
                 }
             } else {
-                key = paste0(case$each, ":", name, "(", each, ")")
+                key <- paste0(case$each, ":", each)
+                if (name != "DEFAULT") {
+                    key <- paste0(key, " - ", name)
+                }
                 newcases[[key]] <- case
             }
         }
@@ -104,41 +113,41 @@ for (name in names(cases)) {
 }
 cases <- newcases
 
-do_enrich = function(expr, odir) {
+do_enrich <- function(expr, odir) {
     print("  Saving expressions ...")
     write.table(
         expr %>% as.data.frame() %>% rownames_to_column("Gene"),
         file.path(odir, "expr.txt"),
-        sep="\t",
-        row.names=TRUE,
-        col.names=TRUE,
-        quote=FALSE
+        sep = "\t",
+        row.names = TRUE,
+        col.names = TRUE,
+        quote = FALSE
     )
     write.table(
         expr %>% as.data.frame() %>% rownames_to_column("Gene") %>% head(n),
         file.path(odir, "exprn.txt"),
-        sep="\t",
-        row.names=TRUE,
-        col.names=TRUE,
-        quote=FALSE
+        sep = "\t",
+        row.names = TRUE,
+        col.names = TRUE,
+        quote = FALSE
     )
 
     print("  Running enrichment ...")
-    enriched = enrichr(rownames(head(expr, n)), dbs)
+    enriched <- enrichr(rownames(head(expr, n)), dbs)  # nolint
     for (db in dbs) {
         write.table(
             enriched[[db]],
             file.path(odir, paste0("Enrichr-", db, ".txt")),
-            sep="\t",
-            row.names=FALSE,
-            col.names=TRUE,
-            quote=FALSE
+            sep = "\t",
+            row.names = FALSE,
+            col.names = TRUE,
+            quote = FALSE
         )
         png(
             file.path(odir, paste0("Enrichr-", db, ".png")),
-            res=100, height=1000, width=1000
+            res = 100, height = 1000, width = 1000
         )
-        print(plotEnrich(enriched[[db]], showTerms = 20, title=db))
+        print(plotEnrich(enriched[[db]], showTerms = 20, title = db))  # nolint
         dev.off()
     }
 }
@@ -146,15 +155,21 @@ do_enrich = function(expr, odir) {
 do_case <- function(casename) {
     print(paste("- Running for case:", casename))
     case <- cases[[casename]]
+    parts <- unlist(strsplit(casename, ":"))
+    section <- parts[1]
+    casename <- paste(parts[-1], collapse = ":")
 
     print("  Calculating average expression ...")
-    avgexpr <- AverageExpression(srtobj, group.by = case$group.by)$RNA[, case$ident, drop = FALSE]
+    avgexpr <- AverageExpression(
+        srtobj,
+        group.by = case$group.by
+    )$RNA[, case$ident, drop = FALSE]
     avgexpr <- avgexpr[order(-avgexpr), , drop = FALSE]
 
-    odir = file.path(outdir, casename)
+    odir <- file.path(outdir, section, casename)
     dir.create(odir, recursive = TRUE, showWarnings = FALSE)
 
     do_enrich(avgexpr, odir)
 }
 
-sapply(names(cases), do_case)
+sapply(sort(names(cases)), do_case)